+++ examples/01-real-estate-proposal/deck.cjs
@@ -15,7 +15,6 @@ module.exports = {
         HERO_LINE_3: '분양성 검토 보고서',
         CONTACT_LABEL: '제출처',
         CONTACT_DETAIL: 'GreenDev Co. 개발사업팀',
-        LOGO_PATH: '',
       },
     },
 
@@ -155,7 +154,7 @@ module.exports = {
     </div>
   </div>
   <div class="logo-area">
-    <img src="../assets/chatdaeri-logo.png" alt="CHATdaeri" style="height:18px;width:auto;display:block">
+    <img src="../assets/logo.png" alt="CHATdaeri" style="height:18px;width:auto;display:block">
   </div>
 </body>
 </html>`,